move readBom methods to Bom class
