+++ dependencies/src/main/java/com/google/cloud/tools/opensource/dependencies/RepositoryUtility.java
@@ -30,14 +30,9 @@ import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Properties;
-import java.util.logging.Logger;
-import org.apache.maven.RepositoryUtils;
 import org.apache.maven.execution.DefaultMavenExecutionRequest;
 import org.apache.maven.execution.MavenExecutionRequest;
-import org.apache.maven.model.DependencyManagement;
 import org.apache.maven.project.MavenProject;
 import org.apache.maven.project.ProjectBuilder;
 import org.apache.maven.project.ProjectBuildingException;
@@ -56,17 +51,12 @@ import org.eclipse.aether.DefaultRepositorySystemSession;
 import org.eclipse.aether.RepositorySystem;
 import org.eclipse.aether.RepositorySystemSession;
 import org.eclipse.aether.artifact.Artifact;
-import org.eclipse.aether.artifact.ArtifactTypeRegistry;
 import org.eclipse.aether.artifact.DefaultArtifact;
 import org.eclipse.aether.collection.DependencySelector;
 import org.eclipse.aether.connector.basic.BasicRepositoryConnectorFactory;
-import org.eclipse.aether.graph.Dependency;
 import org.eclipse.aether.impl.DefaultServiceLocator;
 import org.eclipse.aether.repository.LocalRepository;
 import org.eclipse.aether.repository.RemoteRepository;
-import org.eclipse.aether.resolution.ArtifactDescriptorException;
-import org.eclipse.aether.resolution.ArtifactDescriptorRequest;
-import org.eclipse.aether.resolution.ArtifactDescriptorResult;
 import org.eclipse.aether.resolution.VersionRangeRequest;
 import org.eclipse.aether.resolution.VersionRangeResolutionException;
 import org.eclipse.aether.resolution.VersionRangeResult;
@@ -84,8 +74,6 @@ import org.eclipse.aether.util.graph.transformer.JavaDependencyContextRefiner;
  * Aether initialization.
  */
 public final class RepositoryUtility {
-
-  private static final Logger logger = Logger.getLogger(RepositoryUtility.class.getName());
   
   public static final RemoteRepository CENTRAL =
       new RemoteRepository.Builder("central", "default", "https://repo1.maven.org/maven2/").build();
@@ -183,29 +171,7 @@ public final class RepositoryUtility {
     }
   }
 
-  // TODO arguably this now belongs in the BOM class
-  public static Bom readBom(Path pomFile) throws MavenRepositoryException {
-    RepositorySystem system = RepositoryUtility.newRepositorySystem();
-    RepositorySystemSession session = RepositoryUtility.newSession(system);
-
-    MavenProject mavenProject = createMavenProject(pomFile, session);
-    String coordinates = mavenProject.getGroupId() + ":" + mavenProject.getArtifactId() 
-        + ":" + mavenProject.getVersion();
-    DependencyManagement dependencyManagement = mavenProject.getDependencyManagement();
-    List<org.apache.maven.model.Dependency> dependencies = dependencyManagement.getDependencies();
-
-    ArtifactTypeRegistry registry = session.getArtifactTypeRegistry();
-    ImmutableList<Artifact> artifacts = dependencies.stream()
-        .map(dependency -> RepositoryUtils.toDependency(dependency, registry))
-        .map(Dependency::getArtifact)
-        .filter(artifact -> !Bom.shouldSkipBomMember(artifact))
-        .collect(toImmutableList());
-    
-    Bom bom = new Bom(coordinates, artifacts);
-    return bom;
-  }
-
-  private static MavenProject createMavenProject(Path pomFile, RepositorySystemSession session)
+  static MavenProject createMavenProject(Path pomFile, RepositorySystemSession session)
       throws MavenRepositoryException {
     // MavenCli's way to instantiate PlexusContainer
     ClassWorld classWorld =
@@ -241,56 +207,6 @@ public final class RepositoryUtility {
     } catch (PlexusContainerException | ComponentLookupException | ProjectBuildingException ex) {
       throw new MavenRepositoryException(ex);
     }
-  }
-
-  /**
-   * Parse the dependencyManagement section of an artifact and return the artifacts included there.
-   */
-  public static Bom readBom(String coordinates) throws ArtifactDescriptorException {
-    return readBom(coordinates, ImmutableList.of(CENTRAL.getUrl()));
-  }
-
-  /**
-   * Parse the dependencyManagement section of an artifact and return the artifacts included there.
-   *
-   * @param mavenRepositoryUrls URLs of Maven repositories when resolving the BOM coordinates
-   */
-  public static Bom readBom(String coordinates, List<String> mavenRepositoryUrls)
-      throws ArtifactDescriptorException {
-    Artifact artifact = new DefaultArtifact(coordinates);
-
-    RepositorySystem system = RepositoryUtility.newRepositorySystem();
-    RepositorySystemSession session = RepositoryUtility.newSession(system);
-
-    ArtifactDescriptorRequest request = new ArtifactDescriptorRequest();
-
-    for (String repositoryUrl : mavenRepositoryUrls) {
-      request.addRepository(mavenRepositoryFromUrl(repositoryUrl));
-    }
-
-    request.setArtifact(artifact);
-
-    ArtifactDescriptorResult resolved = system.readArtifactDescriptor(session, request);
-    List<Exception> exceptions = resolved.getExceptions();
-    if (!exceptions.isEmpty()) {
-      throw new ArtifactDescriptorException(resolved, exceptions.get(0).getMessage());
-    }
-    
-    List<Artifact> managedDependencies = new ArrayList<>();
-    for (Dependency dependency : resolved.getManagedDependencies()) {
-      Artifact managed = dependency.getArtifact();
-      if (Bom.shouldSkipBomMember(managed)) {
-        continue;
-      }
-      if (!managedDependencies.contains(managed)) {
-        managedDependencies.add(managed);
-      } else {
-        logger.severe("Duplicate dependency " + dependency);
-      }
-    }
-    
-    Bom bom = new Bom(coordinates, ImmutableList.copyOf(managedDependencies));
-    return bom;
   }
 
   /**
