+++ linkage-monitor/src/test/java/com/google/cloud/tools/dependencies/linkagemonitor/LinkageMonitorTest.java
@@ -93,7 +93,7 @@ public class LinkageMonitorTest {
   public void testBomSnapshot()
       throws ModelBuildingException, ArtifactResolutionException, ArtifactDescriptorException {
 
-    Bom bom = RepositoryUtility.readBom("com.google.cloud:libraries-bom:1.2.0");
+    Bom bom = Bom.readBom("com.google.cloud:libraries-bom:1.2.0");
     Bom snapshotBom =
         LinkageMonitor.copyWithSnapshot(
             system,
